move bundler to packages
