+++ background/provider.ts
@@ -1,7 +1,4 @@
-import {
-  BundlerMode,
-  BundlerProvider
-} from "~packages/provider/bundler/provider"
+import { BundlerMode, BundlerProvider } from "~packages/bundler/provider"
 import { NullUserOperationAuthorizer } from "~packages/provider/waallet/background/authorizer/userOperation/null"
 import { PopUpUserOperationAuthorizer } from "~packages/provider/waallet/background/authorizer/userOperation/popup"
 import { WaalletBackgroundProvider } from "~packages/provider/waallet/background/provider"
