+++ config/test/index.ts
@@ -1,10 +1,7 @@
 import * as ethers from "ethers"
 
-import { UserOperation } from "~packages/provider/bundler"
-import {
-  BundlerMode,
-  BundlerProvider
-} from "~packages/provider/bundler/provider"
+import { UserOperation } from "~packages/bundler"
+import { BundlerMode, BundlerProvider } from "~packages/bundler/provider"
 import type { HexString, Nullable } from "~typing"
 
 class BundlerProviderWithCache extends BundlerProvider {
