+++ packages/provider/waallet/background/authorizer/userOperation/index.ts
@@ -1,4 +1,4 @@
-import { UserOperation } from "~packages/provider/bundler"
+import { UserOperation } from "~packages/bundler"
 
 export type UserOperationAuthorizeCallback = {
   onApproved: (
