+++ packages/provider/waallet/background/provider.ts
@@ -1,10 +1,10 @@
 import * as ethers from "ethers"
 
 import { type Account } from "~packages/account"
+import { UserOperation } from "~packages/bundler"
+import { BundlerProvider } from "~packages/bundler/provider"
 import { type Paymaster } from "~packages/paymaster"
 import { NullPaymaster } from "~packages/paymaster/NullPaymaster"
-import { UserOperation } from "~packages/provider/bundler"
-import { BundlerProvider } from "~packages/provider/bundler/provider"
 import { JsonRpcProvider } from "~packages/provider/jsonrpc/provider"
 import number from "~packages/util/number"
 import type { BigNumberish, HexString } from "~typing"
